requirements.txt
# =============================================================================
# CustomOCR Pipeline API - Production Requirements (Katonic Deployment)
# =============================================================================
# Versions from your running environment's pip freeze. Exact matches.
#
# ⚠️  OpenCV is NOT in this file. See install_deps.sh for why.
#     paddleocr pulls opencv-contrib-python (GUI) which conflicts with
#     the headless variant, AND your env has a conda-installed opencv
#     that pip cannot uninstall. The install script handles this.
#
# INSTALL: Just run the install script:
#     bash install_deps.sh
# =============================================================================

# --- Core API Framework ---
fastapi==0.115.12
starlette==0.46.2
uvicorn==0.34.3
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-multipart==0.0.20
aiofiles==24.1.0

# --- HTTP / Networking ---
anyio==4.11.0
httpx==0.28.1
httpcore==1.0.9
httpx-sse==0.4.3
h11==0.16.0
requests==2.32.5
requests-toolbelt==1.0.0
urllib3==2.6.2
certifi==2025.10.5
idna==3.11
charset-normalizer==3.4.4
chardet==5.2.0
sniffio==1.3.1

# --- Async Networking ---
aiohttp==3.13.2
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
frozenlist==1.8.0
multidict==6.7.0
yarl==1.22.0
propcache==0.4.1
attrs==25.4.0

# --- PDF Processing ---
PyMuPDF==1.26.7
pypdfium2==5.2.0
pdf2image==1.17.0
pikepdf==10.1.0
img2pdf==0.6.3
reportlab==4.4.7

# --- OCR Engine (PaddleOCR) ---
paddlepaddle==3.2.0
paddleocr==3.3.2
paddlex==3.3.12
aistudio-sdk==0.3.8
# NOTE: paddleocr is installed via install_deps.sh with --no-deps
#       to prevent it from pulling opencv-contrib-python (GUI).

# --- Computer Vision ---
# OpenCV handled by install_deps.sh — DO NOT add here.
supervision==0.27.0

# --- Image Processing ---
pillow==11.1.0

# --- Document Parsing ---
python-docx==1.2.0
openpyxl==3.1.5
beautifulsoup4==4.14.3
lxml==6.0.2
soupsieve==2.8.1
et_xmlfile==2.0.0

# --- Arabic / RTL Text ---
arabic-reshaper==3.0.0
python-bidi==0.6.7

# --- NLP / Tokenization ---
tiktoken==0.12.0
sentencepiece==0.2.1
regex==2025.11.3
ftfy==6.3.1

# --- AI / ML Frameworks ---
numpy==1.26.4
scipy==1.16.3
scikit-learn==1.4.2
pandas==2.1.4
pyarrow==14.0.2
einops==0.8.1
safetensors==0.7.0
opt-einsum==3.3.0
joblib==1.5.3
threadpoolctl==3.6.0

# --- LangChain ---
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.81
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.5.0

# --- OpenAI / LLM ---
openai==2.14.0
jiter==0.12.0

# --- HuggingFace ---
huggingface_hub==1.2.3
tokenizers==0.22.1
hf-xet==1.2.0

# --- ModelScope ---
modelscope==1.33.0

# --- Utilities ---
psutil==7.2.0
tqdm==4.67.1
colorlog==6.10.1
python-dotenv==1.2.1
filelock==3.20.1
PyYAML==6.0.2
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
Jinja2==3.1.6
MarkupSafe==3.0.3
click==8.3.1
typing_extensions==4.15.0
packaging==24.1
setuptools==80.9.0
wheel==0.45.1
future==1.0.0
distro==1.9.0
shellingham==1.5.4
uuid_utils==0.12.0

# --- Type Inspection ---
typing-inspect==0.9.0
typing-inspection==0.4.2
mypy_extensions==1.1.0
annotated-types==0.7.0

# --- CLI ---
typer-slim==0.20.1

# --- Data Serialization ---
orjson==3.11.5
ujson==5.11.0
marshmallow==3.26.2
dataclasses-json==0.6.7
jsonpatch==1.33
jsonpointer==3.0.0

# --- XML / HTML ---
cssselect==1.3.0
cssutils==2.11.1
premailer==3.10.0
defusedxml==0.7.1

# --- Geometry / Vision Utilities ---
shapely==2.1.2
pyclipper==1.4.0
networkx==3.5
graphviz==0.21
imagesize==1.4.1
py-cpuinfo==9.0.0

# --- Scientific / Plotting ---
matplotlib==3.10.8
contourpy==1.3.3
fonttools==4.61.1
kiwisolver==1.4.9
cycler==0.12.1
pyparsing==3.3.1

# --- Database ---
SQLAlchemy==2.0.45
greenlet==3.3.0
pymssql==2.3.11

# --- Date / Time ---
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.3

# --- Protobuf ---
protobuf==6.30.2

# --- Misc Transitive Dependencies ---
six==1.17.0
fsspec==2025.12.0
prettytable==3.17.0
tenacity==9.1.2
Deprecated==1.3.1
wrapt==2.0.1
cachetools==6.2.4
more-itertools==10.8.0
colorama==0.4.6
nest-asyncio==1.6.0
decorator==5.2.1
wcwidth==0.2.14

# --- Crypto / Security ---
cryptography==44.0.2
cffi==1.16.0
pycparser==2.22
pycryptodome==3.23.0

# --- Cloud SDK (paddlepaddle dependency) ---
bce-python-sdk==0.9.56

# --- Compression ---
Brotli==1.1.0
brotlicffi==1.1.0.0
zstandard==0.25.0

# ==============================================================================
# NEW packages to ADD to your existing requirements.txt
# Only these 5 lines need to be added — everything else stays the same.
# ==============================================================================
 
# Celery with Redis transport support
celery[redis]>=5.4.0
 
# Redis Python client (Celery dependency, also used directly)
redis>=5.0.0
 
# gevent: green thread pool for I/O-bound Celery workers
# 24 greenlets ≈ 50 MB vs 24 OS threads ≈ 200 MB
gevent>=24.2.1
 
# Flower: Celery monitoring web dashboard (port 5555)
flower>=2.0.0
 
# SQLAlchemy: only needed if POSTGRES_URL is set in .env
# Safe to include even if Postgres is not used — it is never imported unless needed
sqlalchemy>=2.0.0